requirements.txt
asyncpg~=0.28.0
aiogram~=2.25.1
aiolimiter==1.1.0
python-dotenv~=1.0.0
sqlalchemy~=2.0.21
pillow~=10.0.0
apscheduler~=3.10.4
miniopy-async==1.16
redis~=5.0.1
aiofiles~=23.2.1